Onclick Website action mostly implemented
Slide opens a website when appropriatly triggered by an on click action.
Added an event which notifies listeners when teh sequence number is changed.
Embedded webrowser deliberatly closes when you try to advance/go back in the presentation.
TODO:  Add back/forward button for browsing? Remove the scroll bars when not needed.